requirements.txt
aiogram==2.25.1
pydantic==1.10.4
yaml-1.3==0.1.0
PyYAML~=6.0
SQLAlchemy==2.0.3
psycopg2==2.9.5
